third_party/blink/web_tests/external/wpt/css/selectors/parsing/parse-part-expected.txt

This is a testharness.js-based test.
[FAIL] "::part(foo):lang(en)" should be a valid selector
  Failed to execute 'querySelector' on 'Document': '::part(foo):lang(en)' is not a valid selector.
[FAIL] "::part(foo):dir(ltr)" should be a valid selector
  Failed to execute 'querySelector' on 'Document': '::part(foo):dir(ltr)' is not a valid selector.
[FAIL] "::part(foo):is(ul)" should be a valid selector
  assert_equals: CSS.supports() reports the expected value expected false but got true
[FAIL] "::part(foo):is(nav ul)" should be a valid selector
  assert_equals: CSS.supports() reports the expected value expected false but got true
[FAIL] "::part(foo):where(ul)" should be a valid selector
  assert_equals: CSS.supports() reports the expected value expected false but got true
[FAIL] "::part(foo):where(nav ul)" should be a valid selector
  assert_equals: CSS.supports() reports the expected value expected false but got true
Harness: the test ran to completion.
